requirements.txt
fastapi
uvicorn
google-generativeai
pinecone
sentence-transformers
python-dotenv
requests
